+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+Version 2.0.0:
+ * It uses global `Promise` if possible.
+ * `configure()` accepts Promise constructor according to the spec.
+
 Version 1.0.1:
  * window.query -> window.Queue